moved sanity check into its own class
